cyphal, TAP_ESC, gz_bridge updated

+++ src/drivers/cyphal/Cyphal.cpp
@@ -392,7 +392,7 @@ void CyphalNode::sendHeartbeat()
 	}
 }
 
-bool UavcanMixingInterface::updateOutputs(bool stop_motors, int16_t outputs[MAX_ACTUATORS], unsigned num_outputs,
+bool UavcanMixingInterface::updateOutputs(bool stop_motors, float outputs[MAX_ACTUATORS], unsigned num_outputs,
 		unsigned num_control_groups_updated)
 {
 	// Note: This gets called from MixingOutput from within its update() function
