+++ src/modules/simulation/gz_bridge/GZMixingInterfaceESC.hpp
@@ -55,7 +55,7 @@ public:
 		_node_mutex(node_mutex)
 	{}
 
-	bool updateOutputs(bool stop_motors, int16_t outputs[MAX_ACTUATORS],
+	bool updateOutputs(bool stop_motors, float outputs[MAX_ACTUATORS],
 			   unsigned num_outputs, unsigned num_control_groups_updated) override;
 
 	MixingOutput &mixingOutput() { return _mixing_output; }
